+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* `core` – Add `Shrine::Attachment[]` shorthand for `Shrine::Attachment.new` (@janko)
+
 ## 3.0.0.rc (2019-09-28)
 
 * `core` – Add `Storage#delete_prefixed` method for deleting all files in specified directory (@jrochkind)